services/search-service/requirements.txt
# search-service Python dependencies
# Python 3.11+
# ============================================================

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Milvus
pymilvus==2.4.0

# Database
asyncpg==0.29.0
redis[hiredis]==5.0.1

# Kafka
aiokafka==0.10.0

# gRPC (bitmap-filter-service client)
grpcio==1.62.0
grpcio-tools==1.62.0
protobuf==4.25.2

# GPU Inference
onnxruntime-gpu==1.17.1
numpy==1.26.4
Pillow==10.2.0

# Monitoring
prometheus-client==0.20.0
prometheus-fastapi-instrumentator==6.1.0

# Resilience
pybreaker==1.2.0
tenacity==8.2.3

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.1

# Utilities
mmh3==4.1.0
structlog==24.1.0
orjson==3.9.13
httpx==0.27.0

# Health
psutil==5.9.8

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.27.0
testcontainers==3.7.1
